+++ src/fides/api/api/v1/endpoints/privacy_experience_endpoints.py
@@ -212,6 +212,7 @@ def privacy_experience_list(
 
     results: List[PrivacyExperience] = []
     should_unescape: Optional[str] = request.headers.get(UNESCAPE_SAFESTR_HEADER)
+    # Builds TCF Experience Contents once here, in case multiple TCF Experiences are requested
     base_tcf_contents: TCFExperienceContents = get_tcf_contents(db)
 
     for privacy_experience in experience_query.order_by(