+++ src/fides/api/schemas/privacy_preference.py
@@ -87,8 +87,9 @@ class PrivacyPreferencesRequest(FidesSchema):
     browser_identity: Identity
     code: Optional[SafeStr]
     tc_string: Optional[str] = Field(
-        description="If supplied, _preferences fields will be ignored"
-    )
+        description="If supplied, *_preferences fields will be ignored in favor of this TC string. Saves preferences"
+        "against purposes, special features, and vendors."
+    )  # TODO: Not yet implemented
     preferences: conlist(ConsentOptionCreate, max_items=200) = []  # type: ignore
     purpose_preferences: conlist(TCFPurposeSave, max_items=200) = []  # type: ignore
     special_purpose_preferences: conlist(TCFSpecialPurposeSave, max_items=200) = []  # type: ignore